# replit_backup/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BlueBook Worksite Safety API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Site {
  id: number;
  name: string;
  address: string;
  createdAt: string;
}

export interface CreateSiteRequest {
  name: string;
  address: string;
}

export type IncidentIncidentType =
  (typeof IncidentIncidentType)[keyof typeof IncidentIncidentType];

export const IncidentIncidentType = {
  injury: "injury",
  near_miss: "near_miss",
  property_damage: "property_damage",
  environmental: "environmental",
  medical_treatment: "medical_treatment",
  lost_time: "lost_time",
  fatality: "fatality",
} as const;

export type IncidentSeverity =
  (typeof IncidentSeverity)[keyof typeof IncidentSeverity];

export const IncidentSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type IncidentStatus =
  (typeof IncidentStatus)[keyof typeof IncidentStatus];

export const IncidentStatus = {
  open: "open",
  under_investigation: "under_investigation",
  closed: "closed",
} as const;

export interface Incident {
  id: number;
  siteId: number;
  siteName: string;
  title: string;
  description: string;
  incidentType: IncidentIncidentType;
  severity: IncidentSeverity;
  status: IncidentStatus;
  location: string;
  reportedBy: string;
  dateOccurred: string;
  createdAt: string;
  correctiveActions?: string | null;
}

export type CreateIncidentRequestIncidentType =
  (typeof CreateIncidentRequestIncidentType)[keyof typeof CreateIncidentRequestIncidentType];

export const CreateIncidentRequestIncidentType = {
  injury: "injury",
  near_miss: "near_miss",
  property_damage: "property_damage",
  environmental: "environmental",
  medical_treatment: "medical_treatment",
  lost_time: "lost_time",
  fatality: "fatality",
} as const;

export type CreateIncidentRequestSeverity =
  (typeof CreateIncidentRequestSeverity)[keyof typeof CreateIncidentRequestSeverity];

export const CreateIncidentRequestSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface CreateIncidentRequest {
  siteId: number;
  title: string;
  description: string;
  incidentType: CreateIncidentRequestIncidentType;
  severity: CreateIncidentRequestSeverity;
  location: string;
  reportedBy: string;
  dateOccurred: string;
}

export type UpdateIncidentRequestStatus =
  (typeof UpdateIncidentRequestStatus)[keyof typeof UpdateIncidentRequestStatus];

export const UpdateIncidentRequestStatus = {
  open: "open",
  under_investigation: "under_investigation",
  closed: "closed",
} as const;

export interface UpdateIncidentRequest {
  status?: UpdateIncidentRequestStatus;
  correctiveActions?: string;
}

export type HazardHazardType =
  (typeof HazardHazardType)[keyof typeof HazardHazardType];

export const HazardHazardType = {
  physical: "physical",
  chemical: "chemical",
  biological: "biological",
  ergonomic: "ergonomic",
  psychosocial: "psychosocial",
  electrical: "electrical",
  fire: "fire",
  environmental: "environmental",
} as const;

export type HazardRiskLevel =
  (typeof HazardRiskLevel)[keyof typeof HazardRiskLevel];

export const HazardRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  extreme: "extreme",
} as const;

export type HazardStatus = (typeof HazardStatus)[keyof typeof HazardStatus];

export const HazardStatus = {
  open: "open",
  actioned: "actioned",
  closed: "closed",
} as const;

export interface Hazard {
  id: number;
  siteId: number;
  siteName: string;
  title: string;
  description: string;
  hazardType: HazardHazardType;
  /**
   * @minimum 1
   * @maximum 5
   */
  likelihood: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  consequence: number;
  riskRating: number;
  riskLevel: HazardRiskLevel;
  status: HazardStatus;
  location: string;
  reportedBy: string;
  controlMeasures?: string | null;
  createdAt: string;
}

export type CreateHazardRequestHazardType =
  (typeof CreateHazardRequestHazardType)[keyof typeof CreateHazardRequestHazardType];

export const CreateHazardRequestHazardType = {
  physical: "physical",
  chemical: "chemical",
  biological: "biological",
  ergonomic: "ergonomic",
  psychosocial: "psychosocial",
  electrical: "electrical",
  fire: "fire",
  environmental: "environmental",
} as const;

export interface CreateHazardRequest {
  siteId: number;
  title: string;
  description: string;
  hazardType: CreateHazardRequestHazardType;
  /**
   * @minimum 1
   * @maximum 5
   */
  likelihood: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  consequence: number;
  location: string;
  reportedBy: string;
}

export type UpdateHazardRequestStatus =
  (typeof UpdateHazardRequestStatus)[keyof typeof UpdateHazardRequestStatus];

export const UpdateHazardRequestStatus = {
  open: "open",
  actioned: "actioned",
  closed: "closed",
} as const;

export interface UpdateHazardRequest {
  status?: UpdateHazardRequestStatus;
  controlMeasures?: string;
}

export interface NearMiss {
  id: number;
  siteId: number;
  siteName: string;
  title: string;
  description: string;
  potentialConsequence: string;
  location: string;
  reportedBy: string;
  immediateActions?: string | null;
  createdAt: string;
}

export interface CreateNearMissRequest {
  siteId: number;
  title: string;
  description: string;
  potentialConsequence: string;
  location: string;
  reportedBy: string;
  immediateActions?: string;
}

export interface DashboardStats {
  totalIncidents: number;
  openIncidents: number;
  totalHazards: number;
  openHazards: number;
  extremeRiskHazards: number;
  totalNearMisses: number;
  criticalIncidents: number;
  recentIncidents: Incident[];
  recentHazards: Hazard[];
}

export type ListIncidentsParams = {
  siteId?: number;
  status?: ListIncidentsStatus;
};

export type ListIncidentsStatus =
  (typeof ListIncidentsStatus)[keyof typeof ListIncidentsStatus];

export const ListIncidentsStatus = {
  open: "open",
  under_investigation: "under_investigation",
  closed: "closed",
} as const;

export type ListHazardsParams = {
  siteId?: number;
  status?: ListHazardsStatus;
};

export type ListHazardsStatus =
  (typeof ListHazardsStatus)[keyof typeof ListHazardsStatus];

export const ListHazardsStatus = {
  open: "open",
  actioned: "actioned",
  closed: "closed",
} as const;

export type ListNearMissesParams = {
  siteId?: number;
};

export type GetDashboardParams = {
  siteId?: number;
};