# replit_backup/lib/api-zod/src/generated/types/createNearMissRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BlueBook Worksite Safety API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateNearMissRequest {
  siteId: number;
  title: string;
  description: string;
  potentialConsequence: string;
  location: string;
  reportedBy: string;
  immediateActions?: string;
}