# replit_backup/lib/api-zod/src/generated/types/incident.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BlueBook Worksite Safety API
 * OpenAPI spec version: 0.1.0
 */
import type { IncidentIncidentType } from "./incidentIncidentType";
import type { IncidentSeverity } from "./incidentSeverity";
import type { IncidentStatus } from "./incidentStatus";

export interface Incident {
  id: number;
  siteId: number;
  siteName: string;
  title: string;
  description: string;
  incidentType: IncidentIncidentType;
  severity: IncidentSeverity;
  status: IncidentStatus;
  location: string;
  reportedBy: string;
  dateOccurred: Date;
  createdAt: Date;
  correctiveActions?: string | null;
}