# replit_backup/lib/api-zod/src/generated/types/updateHazardRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BlueBook Worksite Safety API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateHazardRequestStatus } from "./updateHazardRequestStatus";

export interface UpdateHazardRequest {
  status?: UpdateHazardRequestStatus;
  controlMeasures?: string;
}